Commit VCR changes to git
This change uses git to commit the result of running http requests
through vcr-archive.

This is a bit of a codedump at the moment as I've been moving things
around and trying to find the best way to split it all up. Now the dust
has settled a bit we can start pulling out testable pieces.