override command method of Browser object
rather than just overriding the visit method override the command method
which is called by everything so we catch all the page fetches.

Because this is also called by things like find we have a list of
methods that we return on straight away rather than trying to save the
current page. Without this scraping becomes quite slow.