Change ScrapedPageArchive to be a class
This makes it more intuitive to use it in a script where you change
between reading from an archive and writing to an archive, instead of
modifying the state globally you just modify it for the instance you're
working with.

I've kept the shorthand for `ScrapedPageArchive.record`, so for the
simple case it will continue to function as before.