Switch to using specific methods for reading from archive
Rather than a global flag to read from the archive this instead uses a
method that returns pages directly from the archive.

Currently this isn't quite working as intended because the
open_from_archive method calls the open method, which in turn calls the
record method, which throws an error when it tries to insert a new
cassette when there's already one in use. I think the solution to this
is to remove the hard dependency on VCR and just use their cassette
format for storage, but read outside of the VCR world.