Fix Ruby 2.0.0 compatibility
This change uses the Ruby 2.0.0 interface for adding headers to an
open-uri response, which is to join them together with ', ' so that
they're in string format, rather than as an array.

We need to do it this way because it wasn't until Ruby 2.1 that open-uri
learned how to deal with headers that have multiple values.