Remove references to page in capybara adapter
This fixes the Capybara adapter to reference local variables rather than
trying to call methods on 'page'. Previously we had this functionality
in Capybara::Poltergeist::Driver, which has a 'page' variable, but we
moved it to Capybara::Poltergeist::Browser, and therefore no longer need
to call methods on 'page' as we can call them directly.