Enable RubyGems deployment from Travis CI
This change means that when Travis notices a new tag that points to a
commit on the `master` branch it will automatically release that tag to
RubyGems as a new version. See the Travis docs [1] for more details.

This configuration uses a `$RUBYGEMS_API_KEY` variable, which I've set
on Travis from the command line using the following command:

    travis env set RUBYGEMS_API_KEY insert_api_key_here

I got the API key from https://rubygems.org/profile/edit, I used an
everypolitician specific RubyGems account that I created, the password
is on the internal mySociety wiki.

[1]: https://docs.travis-ci.com/user/deployment/rubygems